Build makefiles separately for linux and rt linux
Add separate mk files for rt and linux rt. This will help to add
separate targets for linux and rt if needed. Also change makefile
script to take a argument "--rt". If "--rt=1" then makefiles for rt
targets will be added. If argument is not given, or if --rt not
equals to 1, then script for build for linux.

Signed-off-by: Paresh Bhagat <p-bhagat@ti.com>